@@ -4,20 +4,28 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=2.17.0 <3.0.0'
+  flutter: ">=1.16.0"
 
 dependencies:
   flutter:
     sdk: flutter
+
+  permission_handler: any
 
   app_global_config:
     path: ../app_global_config/
     
   app_network:
     path: ../app_network/
+  flutter_network_kit:
+    path: ../flutter_network_kit/
+
   app_environment:
     path: ../app_environment/
+  flutter_environment_base:
+    path: ../flutter_environment_base/
+
   flutter_baseui_kit:
     path: ../flutter_baseui_kit/
   flutter_overlay_kit:
