@@ -4,8 +4,8 @@ version: 0.0.1
 homepage:
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=1.17.0"
+  sdk: '>=2.17.0 <3.0.0'
+  flutter: ">=1.16.0"
 
 dependencies:
   flutter:
@@ -18,10 +18,8 @@ dependencies:
   flutter_network_kit:
     path: ../flutter_network_kit/
 
-  app_environment:
-    path: ../app_environment/
-  flutter_environment:
-    path: ../flutter_environment/ # 为了引入 ApiManager
+  flutter_environment_base:
+    path: ../flutter_environment_base/ # 为了引入 ApiManager
   
   flutter_overlay_kit:
     path: ../flutter_overlay_kit/ # 为了引入 ToastUtil
