@@ -1,18 +1,32 @@
 /*
  * @Author: dvlproad
- * @Date: 2022-04-16 04:32:55
+ * @Date: 2022-08-10 20:10:50
  * @LastEditors: dvlproad
- * @LastEditTime: 2022-04-18 00:00:30
- * @Description: url转模拟地址的方法
+ * @LastEditTime: 2023-03-25 11:56:41
+ * @Description: 
  */
-import 'package:meta/meta.dart'; // 为了使用 @required
-import 'package:flutter_network_kit/flutter_network_kit.dart';
-import 'package:flutter_network_base/src/mock/local_mock_util.dart';
-import 'package:flutter_environment/flutter_environment.dart';
+import 'package:flutter_network_base/flutter_network_base.dart';
+import 'package:flutter_environment_base/flutter_environment_base.dart';
+
+class AppMockManager {
+  static tryDealApi(String api, {required bool isGet}) {
+    if (ApiManager().allowMock == true) {
+      ApiManager.tryAddApi(api, isGet: isGet);
+      bool shouldMock = ApiManager.shouldAfterMockApi(api);
+      if (shouldMock) {
+        api = api.toSimulateApi();
+      }
+    }
+  }
+}
 
 extension SimulateExtension on String {
   String toSimulateApi() {
-    String simulateApiHost = ApiManager.instance.mockApiHost;
+    String? simulateApiHost = ApiManager().mockApiHost;
+    if (simulateApiHost == null) {
+      return this;
+    }
+
     List<String> allMockApiHosts = ["http://dev.api.xxx.com/hapi/"];
     String newApi = this.newApi(
       newApiHost: simulateApiHost,
