@@ -2,7 +2,7 @@
  * @Author: dvlproad
  * @Date: 2022-04-15 22:08:25
  * @LastEditors: dvlproad
- * @LastEditTime: 2023-03-19 13:47:36
+ * @LastEditTime: 2023-03-25 13:45:44
  * @Description: app的开发工具框架
  */
 library app_devtool_framework;
@@ -10,5 +10,8 @@ library app_devtool_framework;
 export './src/init/dev_tool_init.dart';
 export './src/apns_util.dart';
 export './src/dev_page.dart';
-export 'package:flutter_environment/flutter_environment.dart'
+export 'package:flutter_environment_base/flutter_environment_base.dart'
     show PackageNetworkType, PackageTargetType, DeviceInfoPage;
+
+export './src/env_network/app_env_network_util.dart';
+export 'package:flutter_network_kit/flutter_network_kit.dart';
