@@ -6,10 +6,9 @@
  * @Description: 
  */
 import 'package:flutter/material.dart';
-import 'package:flutter/cupertino.dart';
 import './history_version_cell.dart';
 import 'package:flutter_updateversion_kit/flutter_updateversion_kit.dart'
-    show HistoryVersionBean, DevBranchBean;
+    show HistoryVersionBean;
 
 class HistoryVersionPage extends StatefulWidget {
   final String? historyRecordTime;
@@ -46,14 +45,14 @@ class _HistoryVersionPageState extends State<HistoryVersionPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('历史版本记录'),
+        title: const Text('历史版本记录'),
       ),
       body: Column(
         children: [
           Text(
             '记录时间:$_historyRecordTime',
             textAlign: TextAlign.start,
-            style: TextStyle(color: Colors.red),
+            style: const TextStyle(color: Colors.red),
           ),
           Expanded(
             child: ListView.builder(
