@@ -1,12 +1,16 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2023-01-30 11:55:19
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2023-03-24 21:18:24
+ * @Description: 
+ */
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:app_environment/app_environment.dart';
+// import 'package:app_environment/app_environment.dart';
 
 void main() {
   test('adds one to input values', () {
-    final calculator = Calculator();
-    expect(calculator.addOne(2), 3);
-    expect(calculator.addOne(-7), -6);
-    expect(calculator.addOne(0), 1);
+    //
   });
 }
