@@ -2,7 +2,7 @@
  * @Author: dvlproad
  * @Date: 2022-05-18 15:06:49
  * @LastEditors: dvlproad
- * @LastEditTime: 2023-03-17 23:01:15
+ * @LastEditTime: 2023-03-25 13:14:18
  * @Description: 应用层的网络库(含正常请求+埋点请求)
  */
 library app_network;
@@ -18,10 +18,9 @@ export './url/wish_url_path.dart';
 // 正常请求+埋点请求
 export './src/app_network/app_network_manager.dart';
 export './src/app_network/app_network_cache_manager.dart';
+export './src/monitor_network/monitor_network_manager.dart';
 export './src/app_response_model_util.dart';
-export './src/app_api_simulate_util.dart';
-// future
-export './src/app_env_network_util.dart';
+
 // callback
 export './src/app_network_callback_util.dart';
 
